% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardio.mri-package.R
\docType{package}
\name{cardio.mri-package}
\alias{cardio.mri}
\alias{cardio.mri-package}
\title{cardio.mri: "ML Cardiomyopaty (RM)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Cardiomiopathy analyses from structured data and magnetic resonance imaging through an integrated deep learning approach.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Luca Vedovelli \email{luca.vedovelli@ubep.unipd.it} (\href{https://orcid.org/0000-0003-4847-2333}{ORCID}) [contributor]
  \item Dario Gregori \email{dario.gregori@ubep.unipd.it} (\href{https://orcid.org/0000-0001-7906-0580}{ORCID}) [contributor, copyright holder]
}

}
\keyword{internal}
