% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcox_loss.R
\name{logsumexp_masked}
\alias{logsumexp_masked}
\title{Log (sum exp risks at risk)}
\usage{
logsumexp_masked(risk_scores, mask, axis = 0L, keepdims = FALSE)
}
\arguments{
\item{risk_scores}{(2D numerical tensor) single row 2D tensor with
predicted risk scores}

\item{mask}{(2D logical tensor) squared tensor/matrix with both
dimension equal to the number of columns in \link{risk_scores}. Every
row represent a subject, in their column are reported indicators
for subjects at risk at the time-of-event/censoring time for the
corresponding reference subject.}

\item{axis}{(int, default is 0) Python 0-indexed axis representing
the subjects (i.e batches)}

\item{keepdims}{(logical, default is FALSE) would you like to
collapse the result the subject dimension (i.e. across the \link{axis})?}
}
\value{
1D numerical tensor reporting the log of the internal sum of
the partial likelihood equation for the chcox nonlinear model if
\link{keepdims} is FALSE, or a 2D tensor otherwise.
}
\description{
Compute the internal sum of the loss function over the subject at
risk for any give subject/time.
}
\examples{
time <- c(12, 15, 37, 24, 39, 14, 11)
risk_scores <- np$array(array(time, dim = c(length(time), 1)))
mask <- make_riskset(time)
logsumexp_masked(risk_scores, mask) # shape = (7)
logsumexp_masked(risk_scores, mask, keepdims = TRUE) # shape = (1, 7)
}
