% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcox_loss.R
\name{make_riskset}
\alias{make_riskset}
\title{Make risk set matrix}
\usage{
make_riskset(time)
}
\arguments{
\item{time}{(num) vector of subjects' times}
}
\value{
(lgl, 2D array) reporting in cel (i, j) if subject j is in
the risk set of subject i.
}
\description{
Creates the matrix representing the risk sets, i.e., every rows
represent a subject of reference, and every column a target subject,
the mask risk set matrix report TRUE for every target subject which
is in the risk set of reference subjects (i.e., ones wiht greater or
equal reported time (of event/censoring))
}
\examples{
time <- c(12, 15, 37, 24, 39, 14, 11)
.make_riskset(time)
}
