% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcox_loss.R
\name{safe_normalize}
\alias{safe_normalize}
\title{Safe normalize}
\usage{
safe_normalize(x)
}
\arguments{
\item{x}{(2D tensor) single column 2D array, i.e. one row per
subject, and onyl the column for the predicted risk score}
}
\value{
a tensor of the same shape of \link{x}
}
\description{
If there are negative scores, it slide everything to have a minimum
in zero. In fact, only concordance order with risk scores metters.
}
\examples{
ok <- array(1:10, dim = c(10, 1))
ko <- array(-1:8, dim = c(10, 1))

ok_normalized <- safe_normalize(ok)
ko_normalized <- safe_normalize(ko)

identical(ok_normalized, ok)
identical(ko_normalized, ok)
}
